app/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1
gunicorn==20.0.4
pytz==2020.1
sentry-sdk==0.18.0
sqlparse==0.3.1
urllib3==1.25.10
